backend/requirements.txt
# API
fastapi==0.135.3
uvicorn[standard]==0.34.3

# Base de datos
sqlalchemy==2.0.41
alembic==1.18.4
psycopg2-binary==2.9.11

# Autenticación
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20

# Almacenamiento MinIO
minio==7.2.15

# OCR
pytesseract==0.3.13
Pillow==11.2.1
pypdf==5.4.0

# IA / NLP
transformers==4.52.0
torch==2.10.0

# Variables de entorno
python-dotenv==1.1.0

# Validación
pydantic[email]==2.11.5
pydantic-settings==2.13.1
pydantic[email]==2.11.5
email-validator==2.2.0